SPIKE: Use Bond in the gateway
There are two modules, rpc_source and rpc_sink. rpc_source is currently empty.
rpc_sink creates a Bond service (Transport service) and client. In
Module_Create we start the service. In Module_Start we create the client and
use it to create a transport from the service (akin to calling
IoTHubTransport_Create in the device SDK). I originally wanted rpc_source to
create/use the client, but got weird boost errors until I put client and
service in the same source file so they could share the transport and loopback
socket.